"add converter that has extention function"
