+++ src/main/kotlin/datasource/CSVDataSource.kt
@@ -28,34 +28,4 @@ open class CSVDataSource(private var fileName: String = Constant.FILE_NAME): Dat
             installs = appFields[Constant.ColumnIndex.INSTALLS].toLong(),
             requiresAndroid = appFields[Constant.ColumnIndex.REQUIRED_ANDROID].convertToDouble(),)
     }
-
-    private fun String.convertToDouble():Double? = this.split(" ").first().toDoubleOrNull()
-
-    fun String.stringToDate(): Date? {
-        return SimpleDateFormat(Constant.DATE_FORMAT).parse(this)
-    }
-
-    private fun String.megaByteConverter(): Long?{
-        var result = ""
-        for (c in this){
-            result += if (c.isDigit())
-                c
-            else if ( c == '.' && !result.contains("."))
-                c
-            else if ( c == '.' && result.contains("."))
-                return null
-            else break
-        }
-        val value = result.toDoubleOrNull()
-        if (value!= null) {
-            return when(this.last()) {
-                'K','k' ->  (value*Constant.KILO_BYTE).toLong()
-                'M','m' -> (value*Constant.KILO_BYTE.pow(2)).toLong()
-                'G','g' -> (value*Constant.KILO_BYTE.pow(3)).toLong()
-                else -> null
-            }
-        }
-        return null
-    }
-
 }
